put the stiffness_form_builder in a seperate file, density_simulation is now a managable size with single responsability!
